RegistrationHome renamed as OfficeHome
