+++ packages/client/src/App.tsx
@@ -37,7 +37,7 @@ import { ReviewCertificateAction } from '@client/views/PrintCertificate/ReviewCe
 import { VerifyCollector } from '@client/views/PrintCertificate/VerifyCollector'
 import { ApplicationForm } from '@client/views/RegisterForm/ApplicationForm'
 import { ReviewForm } from '@client/views/RegisterForm/ReviewForm'
-import { RegistrationHome } from '@client/views/RegistrationHome/RegistrationHome'
+import { OfficeHome } from '@client/views/OfficeHome/OfficeHome'
 import { SearchResult } from '@client/views/SearchResult/SearchResult'
 import { SelectInformant } from '@client/views/SelectInformant/SelectInformant'
 import { SelectPrimaryApplicant } from '@client/views/SelectPrimaryApplicant/SelectPrimaryApplicant'
@@ -211,12 +211,12 @@ export class App extends React.Component<IAppProps> {
                                           <ProtectedRoute
                                             exact
                                             path={routes.REGISTRAR_HOME}
-                                            component={RegistrationHome}
+                                            component={OfficeHome}
                                           />
                                           <ProtectedRoute
                                             exact
                                             path={routes.REGISTRAR_HOME_TAB}
-                                            component={RegistrationHome}
+                                            component={OfficeHome}
                                           />
                                           <ProtectedRoute
                                             exact
