+++ packages/client/src/SubmissionController.ts
@@ -24,7 +24,7 @@ import { Action } from '@client/forms'
 import { getRegisterForm } from '@client/forms/register/application-selectors'
 import { AppStore } from '@client/store'
 import { getMutationMapping } from '@client/views/DataProvider/MutationProvider'
-import { REGISTRATION_HOME_QUERY } from '@client/views/RegistrationHome/queries'
+import { REGISTRATION_HOME_QUERY } from '@client/views/OfficeHome/queries'
 import { getOperationName } from 'apollo-utilities'
 import { client } from '@client/utils/apolloClient'
 import moment from 'moment'
