+++ packages/client/src/i18n/messages/views/registrarHome.ts
@@ -11,7 +11,7 @@
  */
 import { defineMessages, MessageDescriptor } from 'react-intl'
 
-interface IregistrationHomeMessages
+interface IOfficeHomeMessages
   extends Record<string | number | symbol, MessageDescriptor> {
   certified: MessageDescriptor
   dataTableResults: MessageDescriptor
@@ -33,7 +33,7 @@ interface IregistrationHomeMessages
   waitingForExternalValidation: MessageDescriptor
 }
 
-const messagesToDefine: IregistrationHomeMessages = {
+const messagesToDefine: IOfficeHomeMessages = {
   certified: {
     defaultMessage: 'Certified',
     description: 'Label for registration status certified',
@@ -126,5 +126,4 @@ const messagesToDefine: IregistrationHomeMessages = {
   }
 }
 
-export const messages: IregistrationHomeMessages =
-  defineMessages(messagesToDefine)
+export const messages: IOfficeHomeMessages = defineMessages(messagesToDefine)
