+++ packages/client/src/views/FieldAgentHome/FieldAgentHome.test.tsx
@@ -33,7 +33,7 @@ import {
 import { waitForElement } from '@client/tests/wait-for-element'
 import { FIELD_AGENT_ROLES } from '@client/utils/constants'
 import { FieldAgentHome } from '@client/views/FieldAgentHome/FieldAgentHome'
-import { EVENT_STATUS } from '@client/views/RegistrationHome/RegistrationHome'
+import { EVENT_STATUS } from '@client/views/OfficeHome/OfficeHome'
 import { ReactWrapper } from 'enzyme'
 import { merge } from 'lodash'
 import * as React from 'react'
