+++ packages/client/src/views/OfficeHome/LoadingIndicator.test.tsx
