+++ packages/client/src/views/OfficeHome/tabs/review/reviewTab.test.tsx
@@ -12,31 +12,35 @@
 import {
   DOWNLOAD_STATUS,
   makeApplicationReadyToDownload,
-  storeApplication
+  storeApplication,
+  modifyApplication
 } from '@client/applications'
 import { Action, Event } from '@client/forms'
 import { checkAuth } from '@client/profile/profileActions'
 import { queries } from '@client/profile/queries'
-import { storage } from '@client/storage'
 import { createStore } from '@client/store'
 import {
   createTestComponent,
   createTestComponentWithApolloClient,
   mockUserResponse,
-  resizeWindow,
-  registrationClerkScopeToken
+  resizeWindow
 } from '@client/tests/util'
-import { waitForElement } from '@client/tests/wait-for-element'
+import { waitForElement, waitFor } from '@client/tests/wait-for-element'
 import { createClient } from '@client/utils/apolloClient'
-import { FETCH_REGISTRATION_BY_COMPOSITION } from '@client/views/RegistrationHome/queries'
-import { RegistrationHome } from '@client/views/RegistrationHome/RegistrationHome'
+import { REGISTRATION_HOME_QUERY } from '@client/views/OfficeHome/queries'
+import {
+  RegistrationHome,
+  EVENT_STATUS
+} from '@client/views/OfficeHome/OfficeHome'
+import { Validate } from '@opencrvs/components/lib/icons'
 import { GridTable } from '@opencrvs/components/lib/interface'
+import ApolloClient from 'apollo-client'
 import { ReactWrapper } from 'enzyme'
 import { merge } from 'lodash'
 import moment from 'moment'
 import * as React from 'react'
 import { Store } from 'redux'
-import { RejectTab } from './rejectTab'
+import { ReviewTab } from './reviewTab'
 import {
   GQLBirthEventSearchSet,
   GQLDeathEventSearchSet
@@ -46,9 +50,6 @@ import { formattedDuration } from '@client/utils/date-formatting'
 const registerScopeToken =
   'eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJzY29wZSI6WyJyZWdpc3RlciIsImNlcnRpZnkiLCJkZW1vIl0sImlhdCI6MTU0MjY4ODc3MCwiZXhwIjoxNTQzMjkzNTcwLCJhdWQiOlsib3BlbmNydnM6YXV0aC11c2VyIiwib3BlbmNydnM6dXNlci1tZ250LXVzZXIiLCJvcGVuY3J2czpoZWFydGgtdXNlciIsIm9wZW5jcnZzOmdhdGV3YXktdXNlciIsIm9wZW5jcnZzOm5vdGlmaWNhdGlvbi11c2VyIiwib3BlbmNydnM6d29ya2Zsb3ctdXNlciJdLCJpc3MiOiJvcGVuY3J2czphdXRoLXNlcnZpY2UiLCJzdWIiOiI1YmVhYWY2MDg0ZmRjNDc5MTA3ZjI5OGMifQ.ElQd99Lu7WFX3L_0RecU_Q7-WZClztdNpepo7deNHqzro-Cog4WLN7RW3ZS5PuQtMaiOq1tCb-Fm3h7t4l4KDJgvC11OyT7jD6R2s2OleoRVm3Mcw5LPYuUVHt64lR_moex0x_bCqS72iZmjrjS-fNlnWK5zHfYAjF2PWKceMTGk6wnI9N49f6VwwkinJcwJi6ylsjVkylNbutQZO0qTc7HRP-cBfAzNcKD37FqTRNpVSvHdzQSNcs7oiv3kInDN5aNa2536XSd3H-RiKR9hm9eID9bSIJgFIGzkWRd5jnoYxT70G0t03_mTVnDnqPXDtyI-lmerx24Ost0rQLUNIg'
 const getItem = window.localStorage.getItem as jest.Mock
-
-const mockFetchUserDetails = jest.fn()
-const mockListSyncController = jest.fn()
 
 const nameObj = {
   data: {
@@ -67,11 +68,13 @@ const nameObj = {
   }
 }
 
-const mockUserData = {
+const mockListSyncController = jest.fn()
+
+const mockSearchData = {
   id: 'e302f7c5-ad87-4117-91c1-35eaf2ea7be8',
   type: 'Birth',
   registration: {
-    status: 'REJECTED',
+    status: 'DECLARED',
     contactNumber: '01622688231',
     trackingId: 'BW0UTHR',
     registrationNumber: null,
@@ -138,115 +141,141 @@ const mockUserData = {
   deceasedName: null,
   createdAt: '2018-05-23T14:44:58+02:00'
 }
-const userData: any = []
+const searchData: any = []
 for (let i = 0; i < 14; i++) {
-  userData.push(mockUserData)
+  searchData.push(mockSearchData)
 }
 merge(mockUserResponse, nameObj)
-mockFetchUserDetails.mockReturnValue(mockUserResponse)
-queries.fetchUserDetails = mockFetchUserDetails
 
-storage.getItem = jest.fn()
-storage.setItem = jest.fn()
+const mockReviewTabData = {
+  totalItems: 2,
+  results: [
+    {
+      id: '9a55d213-ad9f-4dcd-9418-340f3a7f6269',
+      type: 'Birth',
+      operationHistories: [
+        {
+          operationType: 'DECLARED',
+          operatedOn: new Date(1544188309380).toISOString(),
+          operatorOfficeName: 'Baniajan Union Parishad',
+          operatorRole: 'FIELD_AGENT',
+          operatorName: [
+            {
+              familyName: 'Al Hasan',
+              firstNames: 'Shakib',
+              use: 'en'
+            },
+            {
+              familyName: null,
+              firstNames: '',
+              use: 'bn'
+            }
+          ],
+          operatorOfficeAlias: ['বানিয়াজান ইউনিয়ন পরিষদ']
+        }
+      ],
+      registration: {
+        status: 'DECLARED',
+        contactNumber: '01622688231',
+        trackingId: 'BW0UTHR',
+        registrationNumber: undefined,
+        eventLocationId: undefined,
+        registeredLocationId: '308c35b4-04f8-4664-83f5-9790e790cde1',
+        createdAt: '1544188309380',
+        modifiedAt: '1544188309380'
+      },
+      dateOfBirth: '2010-10-10',
+      childName: [
+        {
+          firstNames: 'Iliyas',
+          familyName: 'Khan',
+          use: 'en'
+        },
+        {
+          firstNames: 'ইলিয়াস',
+          familyName: 'খান',
+          use: 'bn'
+        }
+      ]
+    } as GQLBirthEventSearchSet,
+    {
+      id: 'bc09200d-0160-43b4-9e2b-5b9e90424e95',
+      type: 'Death',
+      registration: {
+        status: 'VALIDATED',
+        trackingId: 'DW0UTHR',
+        registrationNumber: undefined,
+        eventLocationId: undefined,
+        contactNumber: undefined,
+        duplicates: ['308c35b4-04f8-4664-83f5-9790e790cd33'],
+        registeredLocationId: '308c35b4-04f8-4664-83f5-9790e790cde1',
+        createdAt: '1544188309380',
+        modifiedAt: '1544188309380'
+      },
+      dateOfDeath: '2007-01-01',
+      deceasedName: [
+        {
+          firstNames: 'Iliyas',
+          familyName: 'Khan',
+          use: 'en'
+        },
+        {
+          firstNames: 'ইলিয়াস',
+          familyName: 'খান',
+          use: 'bn'
+        }
+      ]
+    } as GQLDeathEventSearchSet
+  ]
+}
 
-describe('RegistrationHome sent for update tab related tests', () => {
-  const { store, history } = createStore()
-  const client = createClient(store)
+describe('RegistrationHome sent for review tab related tests', () => {
+  let store: ReturnType<typeof createStore>['store']
+  let history: ReturnType<typeof createStore>['history']
+  let client: ApolloClient<{}>
 
-  beforeAll(async () => {
-    getItem.mockReturnValue(registrationClerkScopeToken)
-    await store.dispatch(checkAuth({ '?token': registrationClerkScopeToken }))
+  beforeEach(async () => {
+    ;(queries.fetchUserDetails as jest.Mock).mockReturnValue(mockUserResponse)
+    const createdStore = createStore()
+    store = createdStore.store
+    history = createdStore.history
+
+    client = createClient(store)
+
+    getItem.mockReturnValue(registerScopeToken)
+    await store.dispatch(checkAuth({ '?token': registerScopeToken }))
   })
 
-  it('renders all items returned from graphql query in sent for update tab', async () => {
+  it('renders all items returned from graphql query in ready for review', async () => {
     const TIME_STAMP = '1544188309380'
+    Date.now = jest.fn(() => 1554055200000)
 
     const testComponent = await createTestComponent(
       // @ts-ignore
-      <RejectTab
+      <ReviewTab
         registrarLocationId={'2a83cf14-b959-47f4-8097-f75a75d1867f'}
         queryData={{
-          data: {
-            totalItems: 2,
-            results: [
-              {
-                id: 'e302f7c5-ad87-4117-91c1-35eaf2ea7be8',
-                type: 'Birth',
-                registration: {
-                  status: 'REJECTED',
-                  contactNumber: '01622688231',
-                  trackingId: 'BW0UTHR',
-                  registrationNumber: undefined,
-                  eventLocationId: undefined,
-                  registeredLocationId: '308c35b4-04f8-4664-83f5-9790e790cde1',
-                  duplicates: [null],
-                  createdAt: '2018-05-23T14:44:58+02:00',
-                  modifiedAt: '2018-05-23T14:44:58+02:00'
-                },
-                dateOfBirth: '2010-10-10',
-                childName: [
-                  {
-                    firstNames: 'Iliyas',
-                    familyName: 'Khan',
-                    use: 'en'
-                  },
-                  {
-                    firstNames: 'ইলিয়াস',
-                    familyName: 'খান',
-                    use: 'bn'
-                  }
-                ]
-              } as GQLBirthEventSearchSet,
-              {
-                id: 'bc09200d-0160-43b4-9e2b-5b9e90424e95',
-                type: 'Death',
-                registration: {
-                  status: 'REJECTED',
-                  trackingId: 'DW0UTHR',
-                  registrationNumber: undefined,
-                  eventLocationId: undefined,
-                  contactNumber: '01622688231',
-                  duplicates: ['308c35b4-04f8-4664-83f5-9790e790cd33'],
-                  registeredLocationId: '308c35b4-04f8-4664-83f5-9790e790cde1',
-                  createdAt: TIME_STAMP,
-                  modifiedAt: TIME_STAMP
-                },
-                dateOfDeath: '2007-01-01',
-                deceasedName: [
-                  {
-                    firstNames: 'Iliyas',
-                    familyName: 'Khan',
-                    use: 'en'
-                  },
-                  {
-                    firstNames: 'ইলিয়াস',
-                    familyName: 'খান',
-                    use: 'bn'
-                  }
-                ]
-              } as GQLDeathEventSearchSet
-            ]
-          }
+          data: mockReviewTabData
         }}
       />,
       store
     )
 
-    const table = await waitForElement(testComponent.component, GridTable)
-    const data = table.prop('content')
-    const EXPECTED_DATE_OF_REJECTION = formattedDuration(
-      moment(
-        moment(TIME_STAMP, 'x').format('YYYY-MM-DD HH:mm:ss'),
-        'YYYY-MM-DD HH:mm:ss'
-      )
+    const gridTable = await waitForElement(testComponent.component, GridTable)
+
+    const data = gridTable.prop('content')
+    const EXPECTED_DATE_OF_APPLICATION = formattedDuration(
+      moment(new Date(Number.parseInt(TIME_STAMP)))
     )
 
     expect(data.length).toBe(2)
-    expect(data[1].id).toBe('bc09200d-0160-43b4-9e2b-5b9e90424e95')
-    expect(data[1].contactNumber).toBe('01622688231')
-    expect(data[1].dateOfRejection).toBe(EXPECTED_DATE_OF_REJECTION)
-    expect(data[1].event).toBe('Death')
-    expect(data[1].actions).toBeDefined()
+    expect(data[0].id).toBe('9a55d213-ad9f-4dcd-9418-340f3a7f6269')
+    expect(data[0].eventTimeElapsed).toBe('8 years ago')
+    expect(data[0].applicationTimeElapsed).toBe(EXPECTED_DATE_OF_APPLICATION)
+    expect(data[0].name).toBe('Iliyas Khan')
+    expect(data[0].trackingId).toBe('BW0UTHR')
+    expect(data[0].event).toBe('Birth')
+    expect(data[0].actions).toBeDefined()
   })
 
   it('returns an empty array incase of invalid graphql query response', async () => {
@@ -254,11 +283,11 @@ describe('RegistrationHome sent for update tab related tests', () => {
 
     const testComponent = await createTestComponent(
       // @ts-ignore
-      <RejectTab
+      <ReviewTab
         registrarLocationId={'2a83cf14-b959-47f4-8097-f75a75d1867f'}
         queryData={{
           data: {
-            totalItems: 2,
+            totalItems: 12,
             results: []
           }
         }}
@@ -266,16 +295,17 @@ describe('RegistrationHome sent for update tab related tests', () => {
       store
     )
 
-    const table = await waitForElement(testComponent.component, GridTable)
-
-    const data = table.prop('content')
+    const gridTable = await waitForElement(testComponent.component, GridTable)
+    const data = gridTable.prop('content')
     expect(data.length).toBe(0)
   })
 
-  it('should show pagination bar in sent for update tab if pagination is used and items more than 11', async () => {
+  it('should show pagination bar if pagination is used and items more than 11 in ReviewTab', async () => {
+    Date.now = jest.fn(() => 1554055200000)
+
     const testComponent = await createTestComponent(
       // @ts-ignore
-      <RejectTab
+      <ReviewTab
         registrarLocationId={'2a83cf14-b959-47f4-8097-f75a75d1867f'}
         queryData={{
           data: {
@@ -288,7 +318,12 @@ describe('RegistrationHome sent for update tab related tests', () => {
       store
     )
 
-    await waitForElement(testComponent.component, '#pagination')
+    const pagination = await waitForElement(
+      testComponent.component,
+      '#pagination'
+    )
+
+    expect(pagination.hostNodes()).toHaveLength(1)
 
     testComponent.component
       .find('#pagination button')
@@ -296,11 +331,12 @@ describe('RegistrationHome sent for update tab related tests', () => {
       .hostNodes()
       .simulate('click')
   })
+  it('should show loadmore button if loadmore is used and items more than 11 in ReviewTab', async () => {
+    Date.now = jest.fn(() => 1554055200000)
 
-  it('should show loadmore button in sent for update tab if loadmore is used and items more than 11', async () => {
     const testComponent = await createTestComponent(
       // @ts-ignore
-      <RejectTab
+      <ReviewTab
         registrarLocationId={'2a83cf14-b959-47f4-8097-f75a75d1867f'}
         queryData={{
           data: {
@@ -313,7 +349,12 @@ describe('RegistrationHome sent for update tab related tests', () => {
       store
     )
 
-    await waitForElement(testComponent.component, '#load_more_button')
+    const loadmore = await waitForElement(
+      testComponent.component,
+      '#load_more_button'
+    )
+
+    expect(loadmore.hostNodes()).toHaveLength(1)
 
     testComponent.component
       .find('#load_more_button')
@@ -322,91 +363,12 @@ describe('RegistrationHome sent for update tab related tests', () => {
       .simulate('click')
   })
 
-  it('renders expanded area for required updates', async () => {
+  it('renders expanded area for validated status', async () => {
     Date.now = jest.fn(() => 1554055200000)
-    const graphqlMock = [
-      {
-        request: {
-          query: FETCH_REGISTRATION_BY_COMPOSITION,
-          variables: {
-            id: 'e302f7c5-ad87-4117-91c1-35eaf2ea7be8'
-          }
-        },
-        result: {
-          data: {
-            fetchRegistration: {
-              id: 'e302f7c5-ad87-4117-91c1-35eaf2ea7be8',
-              registration: {
-                id: '345678',
-                type: 'BIRTH',
-                certificates: null,
-                status: [
-                  {
-                    id: '17e9b24-b00f-4a0f-a5a4-9c84c6e64e98/_history/86c3044a-329f-418',
-                    timestamp: '2019-04-03T07:08:24.936Z',
-                    user: {
-                      id: '153f8364-96b3-4b90-8527-bf2ec4a367bd',
-                      name: [
-                        {
-                          use: 'en',
-                          firstNames: 'Mohammad',
-                          familyName: 'Ashraful'
-                        },
-                        {
-                          use: 'bn',
-                          firstNames: '',
-                          familyName: ''
-                        }
-                      ],
-                      role: 'LOCAL_REGISTRAR'
-                    },
-                    location: {
-                      id: '123',
-                      name: 'Kaliganj Union Sub Center',
-                      alias: ['']
-                    },
-                    office: {
-                      id: '123',
-                      name: 'Kaliganj Union Sub Center',
-                      alias: [''],
-                      address: {
-                        district: '7876',
-                        state: 'iuyiuy'
-                      }
-                    },
-                    type: 'REJECTED',
-                    comments: [
-                      {
-                        comment: 'reason=duplicate&comment=dup'
-                      }
-                    ]
-                  }
-                ],
-                contact: 'MOTHER',
-                contactPhoneNumber: '01622688231'
-              },
-              child: {
-                id: 'FAKE_ID',
-                name: [
-                  {
-                    use: 'en',
-                    firstNames: 'Mushraful',
-                    familyName: 'Hoque'
-                  }
-                ],
-                birthDate: '01-01-1984'
-              },
-              deceased: null,
-              informant: null
-            }
-          }
-        }
-      }
-    ]
 
     const testComponent = await createTestComponent(
       // @ts-ignore
-      <RejectTab
+      <ReviewTab
         registrarLocationId={'2a83cf14-b959-47f4-8097-f75a75d1867f'}
         queryData={{
           data: {
@@ -416,7 +378,7 @@ describe('RegistrationHome sent for update tab related tests', () => {
                 id: 'e302f7c5-ad87-4117-91c1-35eaf2ea7be8',
                 type: 'Birth',
                 registration: {
-                  status: 'REJECTED',
+                  status: 'DECLARED',
                   contactNumber: '01622688231',
                   trackingId: 'BW0UTHR',
                   registrationNumber: undefined,
@@ -426,29 +388,6 @@ describe('RegistrationHome sent for update tab related tests', () => {
                   createdAt: '2018-05-23T14:44:58+02:00',
                   modifiedAt: '2018-05-23T14:44:58+02:00'
                 },
-                operationHistories: [
-                  {
-                    operationType: 'REJECTED',
-                    operatedOn: '2019-12-12T15:24:53.586Z',
-                    operatorRole: 'LOCAL_REGISTRAR',
-                    operatorName: [
-                      {
-                        firstNames: 'Mohammad',
-                        familyName: 'Ashraful',
-                        use: 'en'
-                      },
-                      {
-                        firstNames: '',
-                        familyName: null,
-                        use: 'bn'
-                      }
-                    ],
-                    operatorOfficeName: 'Alokbali Union Parishad',
-                    operatorOfficeAlias: ['আলোকবালী  ইউনিয়ন পরিষদ'],
-                    rejectReason: 'missing_supporting_doc',
-                    rejectComment: 'No supporting documents provided.'
-                  }
-                ],
                 dateOfBirth: '2010-10-10',
                 childName: [
                   {
@@ -467,16 +406,37 @@ describe('RegistrationHome sent for update tab related tests', () => {
                 id: 'bc09200d-0160-43b4-9e2b-5b9e90424e95',
                 type: 'Death',
                 registration: {
-                  status: 'REJECTED',
+                  status: 'VALIDATED',
                   trackingId: 'DW0UTHR',
                   registrationNumber: undefined,
                   eventLocationId: undefined,
-                  contactNumber: '01622688231',
+                  contactNumber: undefined,
                   duplicates: ['308c35b4-04f8-4664-83f5-9790e790cd33'],
                   registeredLocationId: '308c35b4-04f8-4664-83f5-9790e790cde1',
                   createdAt: '2007-01-01',
                   modifiedAt: '2007-01-01'
                 },
+                operationHistories: [
+                  {
+                    operationType: 'VALIDATED',
+                    operatedOn: '2019-12-12T15:23:21.280Z',
+                    operatorRole: 'REGISTRATION_AGENT',
+                    operatorName: [
+                      {
+                        firstNames: 'Tamim',
+                        familyName: 'Iqbal',
+                        use: 'en'
+                      },
+                      {
+                        firstNames: '',
+                        familyName: null,
+                        use: 'bn'
+                      }
+                    ],
+                    operatorOfficeName: 'Alokbali Union Parishad',
+                    operatorOfficeAlias: ['আলোকবালী  ইউনিয়ন পরিষদ']
+                  }
+                ],
                 dateOfDeath: '2007-01-01',
                 deceasedName: [
                   {
@@ -495,34 +455,133 @@ describe('RegistrationHome sent for update tab related tests', () => {
           }
         }}
       />,
-      store,
-      graphqlMock
+      store
     )
 
-    // wait for mocked data to load mockedProvider
-    await new Promise((resolve) => {
-      setTimeout(resolve, 200)
-    })
-    testComponent.component.update()
-    const instance = testComponent.component.find(GridTable).instance() as any
+    const gridTable = (
+      await waitForElement(testComponent.component, GridTable)
+    ).instance()
 
-    instance.toggleExpanded('e302f7c5-ad87-4117-91c1-35eaf2ea7be8')
-    // wait for mocked data to load mockedProvider
-    await new Promise((resolve) => {
-      setTimeout(resolve, 100)
-    })
+    gridTable.toggleExpanded('bc09200d-0160-43b4-9e2b-5b9e90424e95')
 
-    testComponent.component.update()
-    expect(testComponent.component.find('#REJECTED-0').hostNodes().length).toBe(
-      1
+    const element = await waitForElement(
+      testComponent.component,
+      '#VALIDATED-0'
     )
+
+    expect(element.hostNodes().length).toBe(1)
+  })
+
+  it('renders expanded area for declared status', async () => {
+    Date.now = jest.fn(() => 1554055200000)
+
+    const testComponent = await createTestComponent(
+      // @ts-ignore
+      <ReviewTab
+        registrarLocationId={'2a83cf14-b959-47f4-8097-f75a75d1867f'}
+        queryData={{
+          data: {
+            totalItems: 2,
+            results: [
+              {
+                id: 'e302f7c5-ad87-4117-91c1-35eaf2ea7be8',
+                type: 'Birth',
+                registration: {
+                  status: 'DECLARED',
+                  contactNumber: '01622688231',
+                  trackingId: 'BW0UTHR',
+                  registrationNumber: undefined,
+                  eventLocationId: undefined,
+                  registeredLocationId: '308c35b4-04f8-4664-83f5-9790e790cde1',
+                  duplicates: [null],
+                  createdAt: '2018-05-23T14:44:58+02:00',
+                  modifiedAt: '2018-05-23T14:44:58+02:00'
+                },
+                dateOfBirth: '2010-10-10',
+                childName: [
+                  {
+                    firstNames: 'Iliyas',
+                    familyName: 'Khan',
+                    use: 'en'
+                  },
+                  {
+                    firstNames: 'ইলিয়াস',
+                    familyName: 'খান',
+                    use: 'bn'
+                  }
+                ]
+              } as GQLBirthEventSearchSet,
+              {
+                id: 'bc09200d-0160-43b4-9e2b-5b9e90424e95',
+                type: 'Death',
+                registration: {
+                  status: 'VALIDATED',
+                  trackingId: 'DW0UTHR',
+                  registrationNumber: undefined,
+                  eventLocationId: undefined,
+                  contactNumber: undefined,
+                  duplicates: ['308c35b4-04f8-4664-83f5-9790e790cd33'],
+                  registeredLocationId: '308c35b4-04f8-4664-83f5-9790e790cde1',
+                  createdAt: '2007-01-01',
+                  modifiedAt: '2007-01-01'
+                },
+                operationHistories: [
+                  {
+                    operationType: 'DECLARED',
+                    operatedOn: '2019-12-12T15:23:21.280Z',
+                    operatorRole: 'FIELD_AGENT',
+                    operatorName: [
+                      {
+                        firstNames: 'Tamim',
+                        familyName: 'Iqbal',
+                        use: 'en'
+                      },
+                      {
+                        firstNames: '',
+                        familyName: null,
+                        use: 'bn'
+                      }
+                    ],
+                    operatorOfficeName: 'Alokbali Union Parishad',
+                    operatorOfficeAlias: ['আলোকবালী  ইউনিয়ন পরিষদ']
+                  }
+                ],
+                dateOfDeath: '2007-01-01',
+                deceasedName: [
+                  {
+                    firstNames: 'Iliyas',
+                    familyName: 'Khan',
+                    use: 'en'
+                  },
+                  {
+                    firstNames: 'ইলিয়াস',
+                    familyName: 'খান',
+                    use: 'bn'
+                  }
+                ]
+              } as GQLDeathEventSearchSet
+            ]
+          }
+        }}
+      />,
+      store
+    )
+
+    const instance = (
+      await waitForElement(testComponent.component, GridTable)
+    ).instance()
+
+    instance.toggleExpanded('bc09200d-0160-43b4-9e2b-5b9e90424e95')
+
+    const element = await waitForElement(testComponent.component, '#DECLARED-0')
+
+    expect(element.hostNodes().length).toBe(1)
   })
 
   describe('handles download status', () => {
     let testComponent: ReactWrapper<{}, {}>
     let createdTestComponent: { component: ReactWrapper; store: Store }
     beforeEach(async () => {
-      const TIME_STAMP = '1544188309380'
       Date.now = jest.fn(() => 1554055200000)
 
       mockListSyncController
@@ -530,78 +589,13 @@ describe('RegistrationHome sent for update tab related tests', () => {
           data: {
             inProgressTab: { totalItems: 0, results: [] },
             notificationTab: { totalItems: 0, results: [] },
-            reviewTab: { totalItems: 0, results: [] },
-            rejectTab: {
-              totalItems: 2,
-              results: [
-                {
-                  id: '9a55d213-ad9f-4dcd-9418-340f3a7f6269',
-                  type: 'Birth',
-                  registration: {
-                    status: 'REJECTED',
-                    contactNumber: '01622688231',
-                    trackingId: 'BW0UTHR',
-                    registrationNumber: null,
-                    eventLocationId: null,
-                    registeredLocationId:
-                      '308c35b4-04f8-4664-83f5-9790e790cde1',
-                    duplicates: null,
-                    createdAt: '2018-05-23T14:44:58+02:00',
-                    modifiedAt: '2018-05-23T14:44:58+02:00'
-                  },
-                  dateOfBirth: '2010-10-10',
-                  childName: [
-                    {
-                      firstNames: 'Iliyas',
-                      familyName: 'Khan',
-                      use: 'en'
-                    },
-                    {
-                      firstNames: 'ইলিয়াস',
-                      familyName: 'খান',
-                      use: 'bn'
-                    }
-                  ],
-                  dateOfDeath: null,
-                  deceasedName: null
-                },
-                {
-                  id: 'bc09200d-0160-43b4-9e2b-5b9e90424e95',
-                  type: 'Death',
-                  registration: {
-                    status: 'REJECTED',
-                    trackingId: 'DW0UTHR',
-                    registrationNumber: null,
-                    eventLocationId: null,
-                    contactNumber: '01622688231',
-                    duplicates: ['308c35b4-04f8-4664-83f5-9790e790cd33'],
-                    registeredLocationId:
-                      '308c35b4-04f8-4664-83f5-9790e790cde1',
-                    createdAt: TIME_STAMP,
-                    modifiedAt: TIME_STAMP
-                  },
-                  dateOfBirth: null,
-                  childName: null,
-                  dateOfDeath: '2007-01-01',
-                  deceasedName: [
-                    {
-                      firstNames: 'Iliyas',
-                      familyName: 'Khan',
-                      use: 'en'
-                    },
-                    {
-                      firstNames: 'ইলিয়াস',
-                      familyName: 'খান',
-                      use: 'bn'
-                    }
-                  ]
-                }
-              ]
-            },
+            reviewTab: mockReviewTabData,
+            rejectTab: { totalItems: 0, results: [] },
             approvalTab: { totalItems: 0, results: [] },
             printTab: { totalItems: 0, results: [] },
             externalValidationTab: { totalItems: 0, results: [] }
-          }
+          },
+          initialSyncDone: true
         })
         .mockReturnValueOnce({
           data: {
@@ -684,15 +678,6 @@ describe('RegistrationHome sent for update tab related tests', () => {
                         comment: 'This is a note'
                       }
                     ],
-                    type: 'REJECTED',
-                    timestamp: null
-                  },
-                  {
-                    comments: [
-                      {
-                        comment: 'This is a note'
-                      }
-                    ],
                     type: 'DECLARED',
                     timestamp: null
                   }
@@ -725,27 +710,15 @@ describe('RegistrationHome sent for update tab related tests', () => {
 
       createdTestComponent = await createTestComponentWithApolloClient(
         // @ts-ignore
-        <RegistrationHome
-          match={{
-            params: {
-              tabId: 'updates'
-            },
-            isExact: true,
-            path: '',
-            url: ''
-          }}
-        />,
+        <RegistrationHome />,
         store,
         client
       )
+
       testComponent = createdTestComponent.component
-      getItem.mockReturnValue(registerScopeToken)
-      await createdTestComponent.store.dispatch(
-        checkAuth({ '?token': registerScopeToken })
-      )
     })
 
-    it('downloads the application after clicking download button', async () => {
+    it('downloads application after clicking download button', async () => {
       const downloadButton = await waitForElement(
         testComponent,
         '#ListItemAction-0-icon'
@@ -766,7 +739,7 @@ describe('RegistrationHome sent for update tab related tests', () => {
 
       const action = await waitForElement(
         testComponent,
-        '#ListItemAction-0-Update'
+        '#ListItemAction-0-Review'
       )
       action.hostNodes().simulate('click')
 
@@ -800,28 +773,14 @@ describe('RegistrationHome sent for update tab related tests', () => {
       expect(errorIcon.hostNodes()).toHaveLength(1)
     })
   })
-})
 
-describe('Tablet tests', () => {
-  const { store } = createStore()
-
-  beforeAll(async () => {
-    getItem.mockReturnValue(registerScopeToken)
-    await store.dispatch(checkAuth({ '?token': registerScopeToken }))
-    resizeWindow(800, 1280)
-  })
-
-  afterEach(() => {
-    resizeWindow(1024, 768)
-  })
-
-  it('redirects to detail page if item is clicked', async () => {
+  it('check the validate icon', async () => {
     const TIME_STAMP = '1544188309380'
     Date.now = jest.fn(() => 1554055200000)
 
     const testComponent = await createTestComponent(
       // @ts-ignore
-      <RejectTab
+      <ReviewTab
         registrarLocationId={'2a83cf14-b959-47f4-8097-f75a75d1867f'}
         queryData={{
           data: {
@@ -831,15 +790,12 @@ describe('Tablet tests', () => {
                 id: 'e302f7c5-ad87-4117-91c1-35eaf2ea7be8',
                 type: 'Birth',
                 registration: {
-                  status: 'REJECTED',
+                  status: 'VALIDATED',
                   contactNumber: '01622688231',
                   trackingId: 'BW0UTHR',
-                  registrationNumber: undefined,
-                  eventLocationId: undefined,
                   registeredLocationId: '308c35b4-04f8-4664-83f5-9790e790cde1',
-                  duplicates: [null],
-                  createdAt: '2018-05-23T14:44:58+02:00',
-                  modifiedAt: '2018-05-23T14:44:58+02:00'
+                  createdAt: TIME_STAMP,
+                  modifiedAt: TIME_STAMP
                 },
                 dateOfBirth: '2010-10-10',
                 childName: [
@@ -859,11 +815,8 @@ describe('Tablet tests', () => {
                 id: 'bc09200d-0160-43b4-9e2b-5b9e90424e95',
                 type: 'Death',
                 registration: {
-                  status: 'REJECTED',
+                  status: 'DECLARED',
                   trackingId: 'DW0UTHR',
-                  registrationNumber: undefined,
-                  eventLocationId: undefined,
-                  contactNumber: '01622688231',
                   duplicates: ['308c35b4-04f8-4664-83f5-9790e790cd33'],
                   registeredLocationId: '308c35b4-04f8-4664-83f5-9790e790cde1',
                   createdAt: TIME_STAMP,
@@ -890,13 +843,216 @@ describe('Tablet tests', () => {
       store
     )
 
-    const element = await waitForElement(testComponent.component, '#row_0')
-    element.hostNodes().simulate('click')
+    const validate = await waitForElement(testComponent.component, Validate)
 
-    await new Promise((resolve) => {
-      setTimeout(resolve, 100)
+    expect(validate).toHaveLength(1)
+  })
+
+  describe.skip('handles download status for possible duplicate application', () => {
+    let testComponent: ReactWrapper<{}, {}>
+    let createdTestComponent: { component: ReactWrapper; store: Store }
+    beforeAll(async () => {
+      Date.now = jest.fn(() => 1554055200000)
+      const graphqlMock = [
+        {
+          request: {
+            query: REGISTRATION_HOME_QUERY,
+            variables: {
+              locationIds: ['2a83cf14-b959-47f4-8097-f75a75d1867f'],
+              count: 10,
+              reviewStatuses: [EVENT_STATUS.DECLARED, EVENT_STATUS.VALIDATED],
+              inProgressSkip: 0,
+              reviewSkip: 0,
+              rejectSkip: 0,
+              approvalSkip: 0,
+              externalValidationSkip: 0,
+              printSkip: 0
+            }
+          },
+          result: {
+            data: {
+              inProgressTab: { totalItems: 0, results: [] },
+              notificationTab: { totalItems: 0, results: [] },
+              reviewTab: mockReviewTabData,
+              rejectTab: { totalItems: 0, results: [] },
+              approvalTab: { totalItems: 0, results: [] },
+              externalValidationTab: { totalItems: 0, results: [] },
+              printTab: { totalItems: 0, results: [] }
+            }
+          }
+        }
+      ]
+
+      createdTestComponent = await createTestComponent(
+        // @ts-ignore
+        <RegistrationHome />,
+        store,
+        graphqlMock
+      )
+
+      getItem.mockReturnValue(registerScopeToken)
+      await createdTestComponent.store.dispatch(
+        checkAuth({ '?token': registerScopeToken })
+      )
+      testComponent = createdTestComponent.component
     })
-    testComponent.component.update()
+
+    it('starts downloading after clicking download button', async () => {
+      const downloadButton = await waitForElement(
+        testComponent,
+        '#ListItemAction-1-icon'
+      )
+
+      downloadButton.hostNodes().simulate('click')
+      testComponent.update()
+
+      expect(
+        testComponent.find('#action-loading-ListItemAction-1').hostNodes()
+      ).toHaveLength(1)
+    })
+
+    it('shows review button when download is complete', async () => {
+      const downloadedApplication = makeApplicationReadyToDownload(
+        Event.DEATH,
+        'bc09200d-0160-43b4-9e2b-5b9e90424e95',
+        Action.LOAD_REVIEW_APPLICATION
+      )
+      downloadedApplication.downloadStatus = DOWNLOAD_STATUS.DOWNLOADED
+      createdTestComponent.store.dispatch(
+        modifyApplication(downloadedApplication)
+      )
+
+      const action = await waitForElement(
+        testComponent,
+        '#ListItemAction-1-Review'
+      )
+
+      expect(action.hostNodes()).toHaveLength(1)
+      action.hostNodes().simulate('click')
+
+      await waitFor(() =>
+        window.location.href.includes(
+          '/duplicates/bc09200d-0160-43b4-9e2b-5b9e90424e95'
+        )
+      )
+    })
+
+    it('shows error when download is failed', async () => {
+      const downloadedApplication = makeApplicationReadyToDownload(
+        Event.DEATH,
+        'bc09200d-0160-43b4-9e2b-5b9e90424e95',
+        Action.LOAD_REVIEW_APPLICATION
+      )
+      downloadedApplication.downloadStatus = DOWNLOAD_STATUS.FAILED
+      createdTestComponent.store.dispatch(
+        modifyApplication(downloadedApplication)
+      )
+
+      testComponent.update()
+
+      const errorIcon = await waitForElement(
+        testComponent,
+        '#action-error-ListItemAction-1'
+      )
+
+      expect(errorIcon.hostNodes()).toHaveLength(1)
+    })
+  })
+})
+
+describe('Tablet tests', () => {
+  const { store } = createStore()
+
+  beforeAll(async () => {
+    resizeWindow(800, 1280)
+  })
+
+  afterEach(() => {
+    resizeWindow(1024, 768)
+  })
+
+  it('redirects to detail page if item is clicked', async () => {
+    const TIME_STAMP = '1544188309380'
+    Date.now = jest.fn(() => 1554055200000)
+
+    const testComponent = await createTestComponent(
+      // @ts-ignore
+      <ReviewTab
+        registrarLocationId={'2a83cf14-b959-47f4-8097-f75a75d1867f'}
+        queryData={{
+          data: {
+            totalItems: 2,
+            results: [
+              {
+                id: 'e302f7c5-ad87-4117-91c1-35eaf2ea7be8',
+                type: 'Birth',
+                registration: {
+                  status: 'VALIDATED',
+                  contactNumber: '01622688231',
+                  trackingId: 'BW0UTHR',
+                  registrationNumber: undefined,
+                  eventLocationId: undefined,
+                  registeredLocationId: '308c35b4-04f8-4664-83f5-9790e790cde1',
+                  duplicates: [null],
+                  createdAt: TIME_STAMP,
+                  modifiedAt: TIME_STAMP
+                },
+                dateOfBirth: '2010-10-10',
+                childName: [
+                  {
+                    firstNames: 'Iliyas',
+                    familyName: 'Khan',
+                    use: 'en'
+                  },
+                  {
+                    firstNames: 'ইলিয়াস',
+                    familyName: 'খান',
+                    use: 'bn'
+                  }
+                ]
+              } as GQLBirthEventSearchSet,
+              {
+                id: 'bc09200d-0160-43b4-9e2b-5b9e90424e95',
+                type: 'Death',
+                registration: {
+                  status: 'DECLARED',
+                  trackingId: 'DW0UTHR',
+                  registrationNumber: undefined,
+                  eventLocationId: undefined,
+                  contactNumber: undefined,
+                  duplicates: ['308c35b4-04f8-4664-83f5-9790e790cd33'],
+                  registeredLocationId: '308c35b4-04f8-4664-83f5-9790e790cde1',
+                  createdAt: TIME_STAMP,
+                  modifiedAt: TIME_STAMP
+                },
+                dateOfDeath: '2007-01-01',
+                deceasedName: [
+                  {
+                    firstNames: 'Iliyas',
+                    familyName: 'Khan',
+                    use: 'en'
+                  },
+                  {
+                    firstNames: 'ইলিয়াস',
+                    familyName: 'খান',
+                    use: 'bn'
+                  }
+                ]
+              } as GQLDeathEventSearchSet
+            ]
+          }
+        }}
+      />,
+      store
+    )
+
+    getItem.mockReturnValue(registerScopeToken)
+    await testComponent.store.dispatch(
+      checkAuth({ '?token': registerScopeToken })
+    )
+
+    const row = await waitForElement(testComponent.component, '#row_0')
+    row.hostNodes().simulate('click')
 
     expect(window.location.href).toContain(
       '/details/e302f7c5-ad87-4117-91c1-35eaf2ea7be8'
