+++ packages/client/src/views/OfficeHome/tabs/review/reviewTab.tsx
@@ -18,8 +18,11 @@ import { REVIEW_EVENT_PARENT_FORM_PAGE } from '@client/navigation/routes'
 import { getScope } from '@client/profile/profileSelectors'
 import { transformData } from '@client/search/transformer'
 import { IStoreState } from '@client/store'
-import { ITheme } from '@client/styledComponents'
+import styled, { ITheme } from '@client/styledComponents'
 import { Scope } from '@client/utils/authUtils'
+import { EVENT_STATUS } from '@client/views/OfficeHome/OfficeHome'
+import { RowHistoryView } from '@client/views/OfficeHome/RowHistoryView'
+import { Duplicate, Validate } from '@opencrvs/components/lib/icons'
 import {
   ColumnContentAlignment,
   GridTable,
@@ -31,11 +34,8 @@ import moment from 'moment'
 import * as React from 'react'
 import { injectIntl, WrappedComponentProps as IntlShapeProps } from 'react-intl'
 import { connect } from 'react-redux'
-import { withTheme } from 'styled-components'
-import { LoadingIndicator } from '@client/views/RegistrationHome/LoadingIndicator'
-import { RowHistoryView } from '@client/views/RegistrationHome/RowHistoryView'
+import ReactTooltip from 'react-tooltip'
 import {
-  buttonMessages,
   constantsMessages,
   dynamicConstantsMessages
 } from '@client/i18n/messages'
@@ -43,9 +43,14 @@ import { messages } from '@client/i18n/messages/views/registrarHome'
 import { IApplication, DOWNLOAD_STATUS } from '@client/applications'
 import { Action } from '@client/forms'
 import { DownloadButton } from '@client/components/interface/DownloadButton'
+import { withTheme } from 'styled-components'
+import { LoadingIndicator } from '@client/views/OfficeHome/LoadingIndicator'
 import { formattedDuration } from '@client/utils/date-formatting'
 
-interface IBaseRejectTabProps {
+const ToolTipContainer = styled.span`
+  text-align: center;
+`
+interface IBaseReviewTabProps {
   theme: ITheme
   scope: Scope | null
   goToPage: typeof goToPage
@@ -63,26 +68,22 @@ interface IBaseRejectTabProps {
   error?: boolean
 }
 
-interface IRejectTabState {
+interface IReviewTabState {
   width: number
 }
 
-type IRejectTabProps = IntlShapeProps & IBaseRejectTabProps
+type IReviewTabProps = IntlShapeProps & IBaseReviewTabProps
 
-class RejectTabComponent extends React.Component<
-  IRejectTabProps,
-  IRejectTabState
+class ReviewTabComponent extends React.Component<
+  IReviewTabProps,
+  IReviewTabState
 > {
   pageSize = 10
-  constructor(props: IRejectTabProps) {
+  constructor(props: IReviewTabProps) {
     super(props)
     this.state = {
       width: window.innerWidth
     }
-  }
-
-  userHasRegisterScope() {
-    return this.props.scope && this.props.scope.includes('register')
   }
 
   componentDidMount() {
@@ -103,56 +104,11 @@ class RejectTabComponent extends React.Component<
       : false
   }
 
-  getColumns = () => {
-    if (this.state.width > this.props.theme.grid.breakpoints.lg) {
-      return [
-        {
-          label: this.props.intl.formatMessage(constantsMessages.type),
-          width: 14,
-          key: 'event'
-        },
-        {
-          label: this.props.intl.formatMessage(constantsMessages.name),
-          width: 23,
-          key: 'name'
-        },
-        {
-          label: this.props.intl.formatMessage(
-            constantsMessages.applicantContactNumber
-          ),
-          width: 21,
-          key: 'contactNumber'
-        },
-        {
-          label: this.props.intl.formatMessage(constantsMessages.sentOn),
-          width: 22,
-          key: 'dateOfRejection'
-        },
-        {
-          label: this.props.intl.formatMessage(messages.listItemAction),
-          width: 20,
-          key: 'actions',
-          isActionColumn: true,
-          alignment: ColumnContentAlignment.CENTER
-        }
-      ]
-    } else {
-      return [
-        {
-          label: this.props.intl.formatMessage(constantsMessages.type),
-          width: 30,
-          key: 'event'
-        },
-        {
-          label: this.props.intl.formatMessage(constantsMessages.name),
-          width: 70,
-          key: 'name'
-        }
-      ]
-    }
+  userHasRegisterScope() {
+    return this.props.scope && this.props.scope.includes('register')
   }
 
-  transformRejectedContent = (data: GQLEventSearchResultSet) => {
+  transformDeclaredContent = (data: GQLEventSearchResultSet) => {
     const { intl } = this.props
     if (!data || !data.results) {
       return []
@@ -165,30 +121,52 @@ class RejectTabComponent extends React.Component<
       )
       const downloadStatus =
         (foundApplication && foundApplication.downloadStatus) || undefined
+      let icon: JSX.Element = <div />
 
-      if (downloadStatus !== DOWNLOAD_STATUS.DOWNLOADED) {
-        actions.push({
-          actionComponent: (
-            <DownloadButton
-              downloadConfigs={{
-                event: reg.event,
-                compositionId: reg.id,
-                action: Action.LOAD_REVIEW_APPLICATION
-              }}
-              key={`DownloadButton-${index}`}
-              status={downloadStatus as DOWNLOAD_STATUS}
-            />
-          )
-        })
-      } else {
-        if (reg.duplicates && reg.duplicates.length > 0) {
+      if (reg.duplicates && reg.duplicates.length > 0) {
+        if (downloadStatus !== DOWNLOAD_STATUS.DOWNLOADED) {
+          actions.push({
+            actionComponent: (
+              <DownloadButton
+                downloadConfigs={{
+                  event: reg.event,
+                  compositionId: reg.id,
+                  action: Action.LOAD_REVIEW_APPLICATION
+                }}
+                key={`DownloadButton-${index}`}
+                status={downloadStatus as DOWNLOAD_STATUS}
+              />
+            )
+          })
+        } else {
           actions.push({
             label: this.props.intl.formatMessage(constantsMessages.review),
             handler: () => this.props.goToReviewDuplicate(reg.id)
           })
+        }
+
+        icon = <Duplicate />
+      } else {
+        if (reg.declarationStatus === EVENT_STATUS.VALIDATED) {
+          icon = <Validate data-tip data-for="validateTooltip" />
+        }
+        if (downloadStatus !== DOWNLOAD_STATUS.DOWNLOADED) {
+          actions.push({
+            actionComponent: (
+              <DownloadButton
+                downloadConfigs={{
+                  event: reg.event,
+                  compositionId: reg.id,
+                  action: Action.LOAD_REVIEW_APPLICATION
+                }}
+                key={`DownloadButton-${index}`}
+                status={downloadStatus as DOWNLOAD_STATUS}
+              />
+            )
+          })
         } else {
           actions.push({
-            label: this.props.intl.formatMessage(buttonMessages.update),
+            label: this.props.intl.formatMessage(constantsMessages.review),
             handler: () =>
               this.props.goToPage(
                 REVIEW_EVENT_PARENT_FORM_PAGE,
@@ -208,16 +186,16 @@ class RejectTabComponent extends React.Component<
       return {
         ...reg,
         event,
-        dateOfRejection:
-          (reg.modifiedAt &&
+        eventTimeElapsed:
+          (reg.dateOfEvent &&
             formattedDuration(
-              moment(
-                moment(reg.modifiedAt, 'x').format('YYYY-MM-DD HH:mm:ss'),
-                'YYYY-MM-DD HH:mm:ss'
-              )
+              moment(reg.dateOfEvent.toString(), 'YYYY-MM-DD')
             )) ||
           '',
+        applicationTimeElapsed:
+          (reg.createdAt && formattedDuration(moment(reg.createdAt))) || '',
         actions,
+        icon,
         rowClickHandler: [
           {
             label: 'rowClickHandler',
@@ -226,6 +204,66 @@ class RejectTabComponent extends React.Component<
         ]
       }
     })
+  }
+
+  getColumns = () => {
+    if (this.state.width > this.props.theme.grid.breakpoints.lg) {
+      return [
+        {
+          label: this.props.intl.formatMessage(constantsMessages.type),
+          width: 14,
+          key: 'event'
+        },
+        {
+          label: this.props.intl.formatMessage(constantsMessages.name),
+          width: 22,
+          key: 'name'
+        },
+        {
+          label: this.props.intl.formatMessage(
+            messages.listItemApplicationDate
+          ),
+          width: 19,
+          key: 'applicationTimeElapsed'
+        },
+        {
+          label: this.props.intl.formatMessage(constantsMessages.eventDate),
+          width: 19,
+          key: 'eventTimeElapsed'
+        },
+        {
+          width: 6,
+          key: 'icons',
+          isIconColumn: true
+        },
+        {
+          label: this.props.intl.formatMessage(messages.listItemAction),
+          width: 20,
+          key: 'actions',
+          isActionColumn: true,
+          alignment: ColumnContentAlignment.CENTER
+        }
+      ]
+    } else {
+      return [
+        {
+          label: this.props.intl.formatMessage(constantsMessages.type),
+          width: 30,
+          key: 'event'
+        },
+        {
+          label: this.props.intl.formatMessage(constantsMessages.name),
+          width: 50,
+          key: 'name'
+        },
+        {
+          width: 20,
+          key: 'icons',
+          isIconColumn: true,
+          alignment: ColumnContentAlignment.RIGHT
+        }
+      ]
+    }
   }
 
   renderExpandedComponent = (itemId: string) => {
@@ -241,8 +279,15 @@ class RejectTabComponent extends React.Component<
 
     return (
       <HomeContent>
+        <ReactTooltip id="validateTooltip">
+          <ToolTipContainer>
+            {this.props.intl.formatMessage(
+              messages.validatedApplicationTooltipForRegistrar
+            )}
+          </ToolTipContainer>
+        </ReactTooltip>
         <GridTable
-          content={this.transformRejectedContent(data)}
+          content={this.transformDeclaredContent(data)}
           columns={this.getColumns()}
           renderExpandedComponent={this.renderExpandedComponent}
           noResultText={intl.formatMessage(constantsMessages.noResults)}
@@ -272,8 +317,8 @@ function mapStateToProps(state: IStoreState) {
   }
 }
 
-export const RejectTab = connect(mapStateToProps, {
+export const ReviewTab = connect(mapStateToProps, {
   goToPage,
   goToReviewDuplicate,
   goToApplicationDetails
-})(injectIntl(withTheme(RejectTabComponent)))
+})(injectIntl(withTheme(ReviewTabComponent)))
